requirements.txt
# Essential dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
gradio>=4.0.0
requests>=2.31.0

# For running Jupyter notebooks
jupyter>=1.0.0
ipython>=8.0.0
